Alien invasion game
The player controls the ship that is in the middle of the bottom
screen edges. The player moves the ship right and left with the cursor keys; 'space' is used to shoot. At the start of the game
the alien fleet is at the top of the screen and gradually descends, also shifting to the side. The player shoots the aliens.
If he manages to shoot down all the aliens, a new fleet appears, which moves faster than the previous one. If the alien collides with the player's ship or reaches the bottom of the screen, the player loses the ship. If the player loses all three ships, the game ends.